add payloads provider and library config

+++ src/consensus/params.h
@@ -81,7 +81,7 @@ struct Params {
     uint256 nMinimumChainWork;
     uint256 defaultAssumeValid;
     // VeriBlock
-    uint64_t VeriBlockPopSecurityHeight;
+    int VeriBlockPopSecurityHeight;
 };
 } // namespace Consensus
